applications/rag/frontend/container/requirements.txt
# Copyright 2024 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
langchain==0.1.9
sentence-transformers==2.5.1
text_generation==0.6.1
google-cloud-dlp==3.12.2
google-cloud-storage==2.9.0
google-cloud-pubsub==2.17.0
google-cloud-datastore==2.15.2
google-cloud-bigquery==3.11.4
google-cloud-language==2.12
google-api-core==2.15
flask_sqlalchemy==3.1.1
google==3.0.0
google-cloud==0.34.0
google-cloud-logging==3.9.0
google-api-python-client==2.114.0
pymysql==1.1.0
cloud-sql-python-connector[pg8000]==1.7.0
langchain-google-cloud-sql-pg==0.4.0
langchain-google-vertexai==0.1.1
langchain-community==0.0.31